Version 1.2.8
15.07.2001 - Removed unnecessary fstat() system call in sf_ftp and sf_scp1.
16.07.2001 - Function handle_options() and trans_exec() put the quotes for
             file names with spaces or ";" at the wrong place. Fixed.
17.07.2001 - sf_ftp failed to reset the ascii buffer. This could sometimes
             lead to files being send in ASCII mode and not in binary.
18.07.2001 - Implemented "do not remove" option for local directories.
20.07.2001 - The directory option "delete unknown files" WRONGLY deleted
             files in the queue as well. For this reason a new option has
             been added "delete queued files", which will delete old
             queued files. The old option works now as documented.
22.07.2001 - Reduce the timeout to a maximum of 20 second for function
             ftp_get_reply().
25.07.2001 - Did not have enough memory in append functions to store
             the binary zero. Fixed.
             lead to a SIGSEGV. Hope this has been fixed.
             are not separated with an empty line. Thanks to Bellucci
             Massimo for reporting this one.
26.07.2001 - Added function ftp_account() required by some FTP proxies.
29.07.2001 - New dialog show_queue to view and delete files from the
             AFD queue.
30.07.2001 - Function init_sf_burst2() failed to remove empty file
             directories. Fixed.
31.07.2001 - In function check_burst_2() forgot to set protocol.
04.08.2001 - Added to new fields to HOST_CONFIG file to store the status
             and some protocol specific options for this host.
05.08.2001 - New function ftp_idle() to set idle time of remote FTP-server.
             code.
07.08.2001 - Fixed some display bugs in job details window of the
             afd_ctrl dialog.

+++ aftp/eval_files.c
@@ -59,8 +59,8 @@ extern int sys_log_fd;
 void
 eval_config_file(char *file_name, struct data *p_db)
 {
-   size_t length;
-   char   *buffer;
+   off_t length;
+   char  *buffer;
 
    if ((length = read_file(file_name, &buffer)) != INCORRECT)
    {
@@ -311,8 +311,8 @@ eval_config_file(char *file_name, struct data *p_db)
 void
 eval_filename_file(char *file_name, struct data *p_db)
 {
-   size_t file_size;
-   char   *buffer;
+   off_t file_size;
+   char  *buffer;
 
    if ((file_size = read_file(file_name, &buffer)) != INCORRECT)
    {
