+++ asmtp/eval_filename_file.c
@@ -56,8 +56,8 @@ extern int sys_log_fd;
 void
 eval_filename_file(char *file_name, struct data *p_db)
 {
-   size_t file_size;
-   char   *buffer;
+   off_t file_size;
+   char  *buffer;
 
    if ((file_size = read_file(file_name, &buffer)) != INCORRECT)
    {
